move pixels below scattering to lookup utils
